+++ src/arch/arm/SConscript
@@ -93,6 +93,7 @@ Source('fs_workload.cc', tags='arm isa')
 Source('regs/int.cc', tags='arm isa')
 Source('regs/misc.cc', tags='arm isa')
 Source('mmu.cc', tags='arm isa')
+Source('mpam.cc', tags='arm isa')
 Source('nativetrace.cc', tags='arm isa')
 Source('pagetable.cc', tags='arm isa')
 Source('pauth_helpers.cc', tags='arm isa')
